pyproject.toml
[project]
name = "ai-security-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.18.0",
    "google-cloud-aiplatform>=1.127.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]
